refactor paging related methods to concern.
